issue #3 반영

+++ todo-app/src/components/ToDoTemplate.js
@@ -4,7 +4,7 @@ import TodoInsert from './TodoInsert';
 import TodoList from './TodoList';
 import './ToDoTemplate.scss';
 
-const ToDoTemplate = ({children}) => {
+const ToDoTemplate = () => {
   const [todos, setTodos] = useState([]);
   const fetchData = async () => {
     try {
@@ -21,6 +21,10 @@ const ToDoTemplate = ({children}) => {
 
   const onInsert = useCallback(
     async (subject) => {
+      if (subject.trim().length === 0) {
+        alert('할 일을 입력해주세요.');
+        return;
+      }
       await axios({
         method: 'post',
         url: 'http://localhost:3001/todos',
@@ -35,6 +39,7 @@ const ToDoTemplate = ({children}) => {
 
   const onRemove = useCallback(
     async (id) => {
+      if (!validCheckId(id)) return;
       await axios.delete(`http://localhost:3001/todos/${id}`);
       await fetchData();
     },
@@ -43,6 +48,7 @@ const ToDoTemplate = ({children}) => {
 
   const onToggle = useCallback(
     async (id) => {
+      if (!validCheckId(id)) return;
       const todo = todos.find((todo) => todo.id === id);
       await axios.put(`http://localhost:3001/todos/${id}/done`, {
         done: !todo.done,
@@ -52,10 +58,17 @@ const ToDoTemplate = ({children}) => {
     [todos],
   );
 
+  const validCheckId = (id) => {
+    if (todos.findIndex((todo) => todo.id === id) === -1) {
+      alert('잘못된 id입니다.');
+      return false;
+    }
+    return true;
+  };
+
   return (
     <div className="TodoTemplate">
       <div className="app-title">일정 관리</div>
-      <div className="content">{children}</div>
       <TodoInsert onInsert={onInsert} />
       <TodoList todos={todos} onRemove={onRemove} onToggle={onToggle} />
     </div>
